Added tests and updated service

+++ src/main/java/ca/mcgill/ecse321/MuseumBackend/dto/TicketRequestDto.java
@@ -30,7 +30,7 @@ public class TicketRequestDto {
   public void setTicketId(int ticketId) {
     this.ticketId = ticketId;
   }
-  public double getTicketId() {
+  public int getTicketId() {
     return ticketId;
   }
